requirements.txt
fastapi
uvicorn
pandas
sklearn.feature_extraction.text
sklearn.metrics.pairwise